`AsyncWebServiceRequestHandler` routes to service.
-- There have been some rumblings of a need to support being able to
choose between one of several different druid clusters. For example,
this may be used to detect bad data and automatically route to a
different cluster with better data. Or it may be used to put two
different druid clusters behind the same API, and route to one or the
other depending on the shape of the query.

-- This PR implements this by introducing a `DruidWebServiceSelector` to
the `AsyncWebServiceRequestHandler`. The `DruidWebServiceSelector` is a
SAM that takes the druid query, api request, and request context and returns
the `DruidWebservice` to send the query to.