Massages the lua configuration.
-- Cleans up the Lua configuration and adds some additional
documentation.

-- This is represents my current thoughts on what the configuration
should look. Note that I haven't touched any of the Java code yet. The
trickiest part remains how to handle metrics. My current thoughts are to
rely on Jackson deserialization. It won't require us to try to hook into
the injection framework somehow, and it does provide flexibility for
building makers. However, things get awkward when you try to inject
behavior (i.e. a function that dynamically selects a ResultSetMapper, or
even a ResultSetMapper). While it can be done, it effectively requires
someone implementing a custom Maker that takes anything other than data
in the constructor to build a constructor that takes a Map<String,
Object> and effectively do their own dependency injection. If injecting
behavior is rare (i.e. restricted almost exclusively to built-in
makers), then this may be acceptable. If it's common, then we'll need to
come up with a different strategy.